Follow-up to 32f81a668f586716fc51fdd9a1d9526a9fc256b3
Also, rather use `static_cast<int>(stuff)` than `int(stuff)`.
See https://cdn.zneix.eu/JjHk5U4.png

+++ src/providers/twitch/TwitchMessageBuilder.cpp
@@ -791,7 +791,8 @@ void TwitchMessageBuilder::runIgnoreReplaces(
                     }
                     twitchEmotes.push_back(TwitchEmoteOccurence{
                         startIndex + pos,
-                        startIndex + pos + emote.first.string.length(),
+                        static_cast<int>(startIndex + pos +
+                                         emote.first.string.length()),
                         emote.second,
                         emote.first,
                     });
